Extracted tests to a separate file
